requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0

# Database dependencies
sqlalchemy==2.0.23

# AI/ML dependencies
google-generativeai==0.3.2

# HTTP client for testing
requests==2.31.0
